% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/weak_sphere.R
\name{besselH}
\alias{besselH}
\title{Hankel function}
\usage{
besselH(k, n, z)

besselH(k, n, z)
}
\arguments{
\item{k}{kind of Hankel function must be 1 or 2}

\item{n}{order of the Hankel funciton}

\item{z}{numeric data vector}

\item{k}{kind of Hankel function must be 1 or 2}

\item{n}{order of the Hankel funciton}

\item{z}{numeric data vector}
}
\description{
Hankel function

Hankel function
}
\examples{
besselH(k=1,n=1,z=1:10)
besselH(k=1,n=1,z=1:10)
}
