% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundvelocity.R
\name{c_Coppens1981}
\alias{c_Coppens1981}
\title{Sound speed according to Coppens et al. (1981)}
\source{
Coppens, A. B. (1981).
Simple equations for the speed of sound in Neptunian waters. The Journal of the Acoustical Society of America, 69(3), 862-863.
\url{http://asa.scitation.org/doi/abs/10.1121/1.385486}
}
\usage{
c_Coppens1981(D, S, T)
}
\arguments{
\item{D}{Depth in meters}

\item{S}{Salinity in parts per thousands}

\item{T}{Temperature in degrees Celsius}
}
\description{
Calculates speed of sound in seawater based on Coppens (1981)
The empirical equation generally holds validity for a temperature range between 0 and 35 degrees Celsius, Salinities between 0 and 45 parts per thousand and a depth range between 0 and 4000 m
}
\examples{
c_Coppens1981(D=100, S=35, T=10)
}
\references{
Coppens, A. B. (1981).
Simple equations for the speed of sound in Neptunian waters. The Journal of the Acoustical Society of America, 69(3), 862-863.
}
