% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundvelocity.R
\name{c_Leroy08}
\alias{c_Leroy08}
\title{Compute speed of sound according to Leroy et al. (2008)}
\source{
Leroy, C. C., Robinson, S. P., & Goldsmith, M. J. (2008).
A new equation for the accurate calculation of sound speed in all oceans. The Journal of the Acoustical Society of America, 124(5), 2774-2782.
\url{http://asa.scitation.org/doi/abs/10.1121/1.2988296}
}
\usage{
c_Leroy08(Z, T, S, lat)
}
\arguments{
\item{Z}{Depth in m}

\item{T}{Temperature in degrees Celsius}

\item{S}{Salinity in parts per thousand}

\item{lat}{Latitude in degrees}
}
\description{
Returns the sound speed according to Leroy et al (2008). This "newer" equation should solve the sound speed within 0.2 m/s for all seas, including the Baltic and Black sea, based on Temperature, Salinity and Latitude. Exceptions are some seas with anomalities close to the bottom. The equation was specifically designed to be used in marine acoustics.
}
\examples{
# TABLE III in Leroy et al. (2008)
# Common oceans, lat = 30°, P= 80 MPa Z= 7808.13 m, S= 34.7\%
lat=30; Z=7808.13; S=34.7; T=c(1,1.5,2,2.5,3)
c_Leroy08(Z,T,S,lat)
# Common oceans, lat = 30°, P= 80 MPa Z= 7808.13 m, T=2 °C
c_Leroy08(Z,T=2,S=seq(33.5,35.5,.5),lat)
# Common oceans, = 30°, P= 5 MPa Z= 497.12 m, S= 35\%
c_Leroy08(Z=497.12,T=seq(-2,20,2),S=35,lat)
# Common oceans, = 30°, P= 5 MPa Z= 497.12 m, T=8 °C
c_Leroy08(Z=497.12,T=8,S=seq(33,37,1),lat)
}
\references{
Leroy, C. C., Robinson, S. P., & Goldsmith, M. J. (2008).
A new equation for the accurate calculation of sound speed in all oceans. The Journal of the Acoustical Society of America, 124(5), 2774-2782.
}
