% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundvelocity.R
\name{c_Mackenzie1981}
\alias{c_Mackenzie1981}
\title{Sound speed according to Mackenzie et al. (1981)}
\source{
Mackenzie, K. V. (1981).
Nine term equation for sound speed in the oceans. The Journal of the Acoustical Society of America, 70(3), 807-812.
\url{http://asa.scitation.org/doi/abs/10.1121/1.386920}
}
\usage{
c_Mackenzie1981(D, S, T)
}
\arguments{
\item{D}{Depth in meters}

\item{S}{Salinity in parts per thousands}

\item{T}{Temperature in degrees Celsius}
}
\description{
Calculate speed of sound in seawater based on MacKenzie (1981)
The empirical equation generally holds validity for a temperature range between 2 and 30 degrees Celsius, Salinities between 25 and 40 parts per thousand and a depth range between 0 and 8000 m
}
\examples{
c_Mackenzie1981(100,35,10)
}
\references{
Mackenzie, K. V. (1981).
Nine term equation for sound speed in the oceans. The Journal of the Acoustical Society of America, 70(3), 807-812.
}
