% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/weak_sphere.R
\name{k}
\alias{k}
\title{Calculate the acoustic wavenumber k}
\usage{
k(c, f)

k(c, f)
}
\arguments{
\item{c}{Soundspeed in m/s}

\item{f}{frequency in Hz (s^-1)}

\item{c}{Soundspeed in m/s}

\item{f}{frequency in Hz (s^-1)}
}
\description{
Calculate the acoustic wavenumber k

Calculate the acoustic wavenumber k
}
\examples{
k(c=1500,f=200000)
k(c=1500,f=200000)
}
