% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{K_p0}
\alias{K_p0}
\title{compressibility at p = 0}
\usage{
K_p0(S, T)
}
\arguments{
\item{S}{Salinity in psu (~ppm)}

\item{T}{Temperature in degrees Celsius}
}
\description{
compressibility at p = 0
}
