% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/weak_sphere.R
\name{LegPoly}
\alias{LegPoly}
\title{Calculate Legendre Polynomial}
\usage{
LegPoly(degree, coeffs)

LegPoly(degree, coeffs)
}
\arguments{
\item{degree}{The degree of the desired polynomial}

\item{coeffs}{the values at which the p[olynomial will be evaluated}

\item{degree}{The degree of the desired polynomial}

\item{coeffs}{the values at which the p[olynomial will be evaluated}
}
\value{
Legendre polynomials

Legendre polynomials
}
\description{
Function to calculate the Legendre Polynomial,
based on the recurrence relation in Numerical Recipes in C (Eq. 4.6.10) with j = j-1
Therefore:
P_j = (2j-1) * P_(j-1) - (j-1) * P_(j-2)
P_0 = 1; P_1 = x

Function to calculate the Legendre Polynomial,
based on the recurrence relation in Numerical Recipes in C (Eq. 4.6.10) with j = j-1
Therefore:
P_j = (2j-1) * P_(j-1) - (j-1) * P_(j-2)
P_0 = 1; P_1 = x
}
\examples{
coeffs=2;degree=3
LegPoly(degree,coeffs)
coeffs=2;degree=3
LegPoly(degree,coeffs)
}
