% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{rho}
\alias{rho}
\title{Seawater Density according to UNESCO formula}
\usage{
rho(S, T, p = 0)
}
\arguments{
\item{S}{Salinity in psu (which is +/- equal to ppm)}

\item{T}{Temperature in degrees}

\item{p}{pressure in Bar}
}
\description{
UNESCO (1981) Tenth report of the joint panel on
oceanographic tables and standards. UNESCO Technical
Papers in Marine Science, Paris, 25p
}
\examples{
rho(S=35,T=0.5,p=10)
rho(8,10) #Should be 1005.94659
}
