% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{rho_p0}
\alias{rho_p0}
\title{Density (rho) at normal atmospheric pressure (p = 0)}
\usage{
rho_p0(S, T)
}
\arguments{
\item{S}{Salinity in psu (~ppm)}

\item{T}{Temperature in degrees Celsius}
}
\description{
Density (rho) at normal atmospheric pressure (p = 0)
}
