% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/weak_sphere.R
\name{Sbessj}
\alias{Sbessj}
\title{Spherical bessel function first kind}
\usage{
Sbessj(x, order)

Sbessj(x, order)
}
\arguments{
\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}

\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}
}
\description{
Spherical bessel function first kind

Spherical bessel function first kind
}
\examples{
Sbessj(x=1:10,order=1)
Sbessj(x=1:10,order=1)
}
