% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/weak_sphere.R
\name{Sbessj_p}
\alias{Sbessj_p}
\title{Derivative of the Spherical bessel function first kind}
\usage{
Sbessj_p(x, order)

Sbessj_p(x, order)
}
\arguments{
\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}

\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}
}
\description{
Derivative of the Spherical bessel function first kind

Derivative of the Spherical bessel function first kind
}
\examples{
Sbessj_p(x=1:10,order=1)
Sbessj_p(x=1:10,order=0)
Sbessj_p(x=1:10,order=1)
Sbessj_p(x=1:10,order=0)
}
