% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weak_sphere.R
\name{SHankel}
\alias{SHankel}
\title{Spherical Hankel function}
\usage{
SHankel(z, order)
}
\arguments{
\item{z}{numeric data vector}

\item{order}{order of the bessel funciton}
}
\description{
Spherical Hankel function
}
\examples{
SHankel(x=1:10,order=1)
SHankel(x=1:10,order=0)
}
