% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/weak_sphere.R
\name{Sneum}
\alias{Sneum}
\title{Spherical bessel function second kind (Neumann)}
\usage{
Sneum(x, order)

Sneum(x, order)
}
\arguments{
\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}

\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}
}
\description{
Spherical bessel function second kind (Neumann)

Spherical bessel function second kind (Neumann)
}
\examples{
Sneum(x=1:10,order=1)
Sneum(x=1:10,order=1)
}
