% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weak_sphere.R
\name{Sneum_p}
\alias{Sneum_p}
\title{Derivative of the Spherical bessel function second kind}
\usage{
Sneum_p(x, order)
}
\arguments{
\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}
}
\description{
Derivative of the Spherical bessel function second kind
}
\examples{
Sneum_p(x=1:10,order=1)
Sneum_p(x=1:10,order=0)
}
