% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weak_sphere.R
\name{TS.sphere2}
\alias{TS.sphere2}
\title{Calculate the analytical solution for a weakly scattering sphere}
\usage{
TS.sphere2(f, r, a, c, h, g, rho)
}
\arguments{
\item{f}{Frequency in Hz (s^-1)}

\item{r}{Range in m from center of sphere}

\item{a}{Radius of sphere}

\item{c}{Soundspeed in surrounding fluid m/s}

\item{h}{Soundspeed contrast inside/surrounding fluid}

\item{g}{Density contrast inside/surrounding fluid}

\item{rho}{Density of surrounding fluid}
}
\description{
Scattered pressure from an incident plane wave upon a fluid sphere. Based on:
Jech, J. M., Horne, J. K., Chu, D., Demer, D. A., Francis, D. T., Gorska, N., ... & Reeder, D. B. (2015).
"Comparisons among ten models of acoustic backscattering used in aquatic ecosystem research."
The Journal of the Acoustical Society of America, 138(6), 3742-3764.
modified from:
Anderson, V. C. (1950). "Sound scattering from a fluid sphere."
The Journal of the Acoustical Society of America, 22(4), 426-431.
}
\examples{
fs <- as.list(seq(10,400, by=1)*1000) #Frequencies
r <- 10 #range
a <- 0.01 # radius
c <- 1477.4 #soundspeed surrounding fluid
rho <- 1026.8 #density surrounding fluid
g <- 1028.9/rho #density contrast
h <- 1480.3/c #soundspeed contrast
TS <- sapply(fs,TS.sphere2,r=r,a=a,c=c,h=h,g=g,rho=rho)
plot(fs,TS, type="l", xlab="Frequency [Hz]",ylab="TS [dB re m2]")
}
