% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weak_sphere.R
\name{lambda}
\alias{lambda}
\title{Calculate the acoustic wave length lambda}
\usage{
lambda(c, f)
}
\arguments{
\item{c}{Soundspeed in m/s}

\item{f}{frequency in Hz (s^-1)}
}
\description{
Calculate the acoustic wave length lambda
}
\examples{
lambda(c=1500,f=200000)
}
