% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weak_sphere.R
\name{SHankel_p}
\alias{SHankel_p}
\title{Derivative of the Spherical Hankel function}
\usage{
SHankel_p(x, order)
}
\arguments{
\item{x}{numeric data vector}

\item{order}{order of the bessel funciton}
}
\description{
Derivative of the Spherical Hankel function
}
\examples{
SHankel_p(x=1:10,order=1)
SHankel_p(x=1:10,order=0)
}
